Added model loading and rendering.
Animations seem to be mostly correct for pilot.ae, with the exception of some objects being angled a bit incorrectly. (such as the drill in Activity_Drill).

Sped up tree creation in the debug archive viewer as well as config file parsing, but not by much. It still takes a couple of seconds to launch the game.